requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
anthropic==0.84.0
pdfplumber==0.11.8
python-docx==1.2.0
openpyxl==3.1.2
openai-whisper==20240930
pytesseract==0.3.13
jinja2==3.1.3
aiofiles==23.2.1
python-slugify==8.0.4
